Break up a bunch of grouped panel text, a mess of broken panel imports
