+++ client/src/panels/panel_declarations/finances/budget_measures/budget_measures_panel.scss
@@ -1,4 +1,4 @@
-@import '../../../common_css/common-variables';
+@import '../../../../common_css/common-variables';
 
 .budget-panel-controls-and-graph {
   width: 100%;
