+++ client/src/panels/panel_declarations/finances/vote_stat/vote_stat_text_provider.js
@@ -1,7 +1,6 @@
 import { 
   create_text_maker_component,
 } from '../../shared.js';
-import text1 from './vote_stat_text.yaml';
-import text2 from '../../../common_text/common_lang.yaml';
+import text from './vote_stat_text.yaml';
 
-export const { text_maker, TM } = create_text_maker_component([text1, text2]);
+export const { text_maker, TM } = create_text_maker_component([text]);
