+++ client/src/panels/panel_declarations/finances/welcome_mat/welcome-mat.scss
@@ -1,4 +1,4 @@
-@import '../../../common_css/common-variables';
+@import '../../../../common_css/common-variables';
 
 /* 
   The parts of the old grid-system.css, as used by welcome-mat, rewritten in BEM and SASS
