+++ client/src/panels/panel_declarations/intro/index.js
@@ -7,4 +7,4 @@ export {
   declare_description_panel,
 } from './profile_panels.js';
 
-export { declare_simplographic_panel } from './simplographic.js';
+export { declare_simplographic_panel } from './simplographic/simplographic.js';
