--- client/src/panels/panel_declarations/intro/intro_lang.yaml
@@ -1,186 +1,0 @@
-# Dept related panels 
-portfolio_structure_intro_title:
-  en: Where does this organization fit in the government structure?
-  fr: Où s'inscrit cette organisation dans la structure gouvernementale?
-
-dept_related_tags_title:
-  en: Tags for this organization
-  fr: Étiquettes pour cette organisation
-
-portfolio_structure_related_title:
-  en: Program Structure
-  fr: Structure de Programmes
-
-dept_is_tagged_with_following:
-  transform: [handlebars,markdown]
-  en: |
-    Programs under **{{subj_name subject}}** have been {{gl_tt "tagged" "TAG"}} under the following areas of interest
-  fr: |
-    Les Programmes sous **{{subj_name subject}}** ont été {{gl_tt "étiquetés" "TAG"}} comme relevant des sujets suivants
-
-
-# Program panel text 
-program_is_tagged_with_following:
-  transform: [handlebars,markdown]
-  en: |
-    **{{subj_name subject}}** has been {{gl_tt "tagged" "TAG"}} with the following areas of interest
-  fr: |
-    Le Programme **{{subj_name subject}}** a été {{gl_tt "étiqueté" "TAG"}} aux domaines d'intérêt suivants
-
-program_desc_title:
-  en: Program Description
-  fr: Description du Programme 
-
-program_fed_structure_title:
-  en: Where does this Program fit in the government structure?
-  fr: Où s'inscrit ce Programme dans la structure gouvernementale?
-
-program_tags_title:
-  en: Tags for this Program
-  fr: Étiquettes pour ce programme
-
-related_program_structure_title:
-  en: Programs under the same organization
-  fr: Programmes sous la même organisation
-
-dead_program_warning_title:
-  en: This Program is no longer active
-  fr: Ce programme n'est plus actif
-
-dead_crso_warning:
-  en: This Core Responsibility is no longer active
-  fr: Cette responsabilité essentielle n'est plus active
-
-dead_program_warning_text:
-  transform: [handlebars, markdown]
-  en: |
-    Unless you're here for historical financial data, select the 'Where can I go from here?' topic above to see all of **{{subj_name subject.dept}}**'s Programs.
-  fr: |
-    À moins que vous voulez des donées financières historiques, sélectionnez le sujet «Où puis-je aller à partir d'ici?» ci-dessus pour voir tous les Programmes **{{de_dept subject.dept}}**.
-
-dead_program_warning:
-  transform: [handlebars, markdown]
-  en: This Program is no longer active; only limited information is available
-  fr: Ce programme n'est plus actif; les renseignements disponibles sont limités
-
-
-
-# Tag text 
-intro_to_graphs_title:
-  en: Introduction to Program Tags
-  fr: Introduction aux étiquettes de Programmes
-
-tag_desc_title:
-  en: Tag Description
-  fr: Description de l'étiquette
-
-tag_fed_structure_title:
-  en: Where does this tag fit in the government tagging scheme?
-  fr: Où s'inscrit cette étiquette dans le système d'étiquetage gouvernemental?
-
-sibling_tags_title:
-  en: List of all the tags included in the same tagging category 
-  fr: Liste des étiquettes de la même catégorie
-
-tag_progs_by_dept_title:
-  en: Which organizations have programs under this tag? 
-  fr: Quelles sont les organisations qui ont des programmes associés à cette étiquette?
-
-m2m_warning_title:
-  en: |
-    How to use tags?
-  fr: |
-    Comment utiliser les étiquettes?
-
-
-# Hierarchy text
-hierarchy_dead_element_icon_alt_text:
-  transform: [handlebars]
-  en:  |
-    This item is either 1) an organization for which the InfoBase has no data, or 2) a Strategic Outcome or program that has ended or merged into a new one between {{pa_last_year_5}} and {{planning_year_1}}.
-  fr: |
-    Cet article est soit 1) une organisation pour lesquels l'InfoBase n'a pas de données, ou 2) un résultat stratégique ou programme qui a pris fin ou qui ont fusionné entre {{pa_last_year_5}} et {{planning_year_1}}
-
-hierarchy_contains_dead_elements:
-  transform: [handlebars]
-  en:  |
-    : Corresponds to Programs and/or Strategic Outcomes that either ended or merged into new ones between {{pa_last_year_5}} and {{planning_year_1}}.
-  fr: |
-    : Corresponds aux programmes ou résultats stratégiques qui ont pris fin ou qui ont fusionné entre {{pa_last_year_5}} et {{planning_year_1}}.
-
-hierarchy_contains_elements_with_limited_data:
-  en:  |
-    : Corresponds to organizations for which the InfoBase has no data.
-  fr: |
-    : Corresponds aux organisations pour lesquels l'InfoBase n'a pas de données.
-
-
-
-# CRSO text
-crso_desc_title:
-  transform: [handlebars]
-  en: |
-    Core Responsibility Description
-  fr: |
-    Description de la responsabilité essentielle
-
-crso_links_to_other_crso_title:
-  transform: [handlebars]
-  en: List of Core Responsibilities and Programs under the same organization
-  fr: Liste de responsabilités essentielles et de programmes sous la même organizations
-
-crso_links_to_prg_title:
-  transform: [handlebars]
-  en: Programs under this Core Responsibility
-  fr: Programmes sous cette responsabilité essentielle
-
-crso_is_tagged_with_following:
-  transform: [handlebars,markdown]
-  en: |
-    Programs under **{{subj_name subject}}** have been {{gl_tt "tagged" "TAG"}} under the following areas of interest
-  fr: |
-    Les Programmes sous **{{subj_name subject}}** ont été {{gl_tt "étiquetés" "TAG"}} comme relevant des sujets suivants
-
-crso_tags_title:
-  en: Tags for this Core Responsibility
-  fr: Étiquettes pour cette responsabilité essentielle 
-
-crso_in_gov_title:
-  transform: [handlebars]
-  en: Where does this Core Responsibility fit in the government reporting structure ?
-  fr: Où s'inscrit cette responsabilité essentielle dans la structure de rapport du gouvernement ?
-
-
-# Late banners
-late_dps_warning_dept:
-  en: The 2019-20 Departmental Plan data is not available for this organization. Information will be updated as soon as the data is submitted.
-  fr: L'information des plans ministériel 2019-2020 n’est pas disponible pour cette organisation. L’information sera mise-à-jour dès que les données auront été soumises.
-
-late_dps_warning_program:
-  en: The 2019-20 Departmental Plan data is not available for this Program. Information will be updated as soon as the data is submitted.
-  fr: L'information des plans ministériel 2019-2020 n’est pas disponible pour ce programme. L’information sera mise-à-jour dès que les données auront été soumises.
-
-late_dps_warning_crso:
-  en: The 2019-20 Departmental Plan data is not available for this Core Responsibility. Information will be updated as soon as the data is submitted.
-  fr: L'information des plans ministériel 2019-2020 n’est pas disponible pour cette responsabilité essentielle. L’information sera mise-à-jour dès que les données auront été soumises.
-
-late_dps_warning_gov:
-  transform: [handlebars]
-  en: The count of results and indicators does not include indicators for {{array_to_grammatical_list late_dp_department_names}}. Information will be updated as soon as the data is submitted.
-  fr: Le nombre de résultats et d’indicateurs n’inclus pas les indicateurs pour {{array_to_grammatical_list late_dp_department_names}}. L’information sera mise-à-jour dès que les données auront été soumises.
-
-
-# Misc.
-links_to_rpb_title:
-  en: Detailed reports by subject
-  fr: Rapports détaillés par sujets
-
-links_to_rpb_text:
-  transform: [ handlebars, markdown ]
-  en: |
-    Select from one of the following tables to see the data behind the infographic and build your own report. 
-  fr: |
-    Choisissez parmi l'un des tableaux ci-dessous pour voir les données qui supportent les infographies et construire votre propre rapport.
-additional_links:
-  en: Additional links
-  fr: Des liens vers l'information supplémentaire
