+++ client/src/panels/panel_declarations/intro/profile_panels.yaml
@@ -1,0 +1,11 @@
+program_desc_title:
+  en: Program Description
+  fr: Description du Programme 
+
+tag_desc_title:
+  en: Tag Description
+  fr: Description de l'étiquette
+
+crso_desc_title:
+  en: Core Responsibility Description
+  fr: Description de la responsabilité essentielle
