+++ client/src/panels/panel_declarations/intro/simplographic/simplographic.js
@@ -1,16 +1,15 @@
 import './simplographic.scss';
 import simplographic_text from './simplographic.yaml';
 
-
 import { 
   Subject, 
   declare_panel, 
   util_components, 
   InfographicPanel,
-} from '../shared.js';
-import { infograph_href_template, rpb_link } from '../../../link_utils.js';
-import { ResultCounts, current_drr_key } from '../../../models/results.js';
-import { get_static_url } from '../../../request_utils.js';
+} from '../../shared.js';
+import { infograph_href_template, rpb_link } from '../../../../link_utils.js';
+import { ResultCounts, current_drr_key } from '../../../../models/results.js';
+import { get_static_url } from '../../../../request_utils.js';
 
 const { Gov, Dept } = Subject;
 const { create_text_maker_component} = util_components;
