+++ client/src/panels/panel_declarations/intro/simplographic/simplographic.scss
@@ -1,4 +1,4 @@
-@import '../../../common_css/common-variables';
+@import '../../../../common_css/common-variables';
 
 @media (max-width: $maxLargeDevice){
   .about-government-intro-grid {
