+++ client/src/panels/panel_declarations/intro/simplographic/simplographic.yaml
