+++ client/src/panels/panel_declarations/misc/hierarchy_panels.yaml
@@ -1,0 +1,49 @@
+
+portfolio_structure_intro_title:
+  en: Where does this organization fit in the government structure?
+  fr: Où s'inscrit cette organisation dans la structure gouvernementale?
+portfolio_structure_related_title:
+  en: Program Structure
+  fr: Structure de Programmes
+
+program_fed_structure_title:
+  en: Where does this Program fit in the government structure?
+  fr: Où s'inscrit ce Programme dans la structure gouvernementale?
+tag_fed_structure_title:
+  en: Where does this tag fit in the government tagging scheme?
+  fr: Où s'inscrit cette étiquette dans le système d'étiquetage gouvernemental?
+
+related_program_structure_title:
+  en: Programs under the same organization
+  fr: Programmes sous la même organisation
+
+sibling_tags_title:
+  en: List of all the tags included in the same tagging category 
+  fr: Liste des étiquettes de la même catégorie
+
+crso_links_to_other_crso_title:
+  transform: [handlebars]
+  en: List of Core Responsibilities and Programs under the same organization
+  fr: Liste de responsabilités essentielles et de programmes sous la même organizations
+crso_in_gov_title:
+  transform: [handlebars]
+  en: Where does this Core Responsibility fit in the government reporting structure?
+  fr: Où s'inscrit cette responsabilité essentielle dans la structure de rapport du gouvernement?
+
+hierarchy_dead_element_icon_alt_text:
+  transform: [handlebars]
+  en:  |
+    This item is either 1) an organization for which the InfoBase has no data, or 2) a Strategic Outcome or program that has ended or merged into a new one between {{pa_last_year_5}} and {{planning_year_1}}.
+  fr: |
+    Cet article est soit 1) une organisation pour lesquels l'InfoBase n'a pas de données, ou 2) un résultat stratégique ou programme qui a pris fin ou qui ont fusionné entre {{pa_last_year_5}} et {{planning_year_1}}
+hierarchy_contains_elements_with_limited_data:
+  en:  |
+    : Corresponds to organizations for which the InfoBase has no data.
+  fr: |
+    : Corresponds aux organisations pour lesquels l'InfoBase n'a pas de données.
+hierarchy_contains_dead_elements:
+  transform: [handlebars]
+  en:  |
+    : Corresponds to Programs and/or Strategic Outcomes that either ended or merged into new ones between {{pa_last_year_5}} and {{planning_year_1}}.
+  fr: |
+    : Corresponds aux programmes ou résultats stratégiques qui ont pris fin ou qui ont fusionné entre {{pa_last_year_5}} et {{planning_year_1}}.
