+++ client/src/panels/panel_declarations/misc/index.js
@@ -18,15 +18,13 @@ export {
   declare_crso_links_to_other_crso_panel,
 } from './hierarchy_panels.js';
 
-export { declare_late_dps_warning_panel } from './late_dps_warning.js';
-
-export { declare_m2m_tag_warning_panel } from '../misc/m2m_tag_warning.js/index.js';
-
 export {
   declare_year_warning_panel,
   declare_dead_program_warning_panel,
   declare_dead_crso_warning_panel,
-} from './meta_panels.js';
+  declare_m2m_tag_warning_panel,
+  declare_late_dps_warning_panel,
+} from './warning_panels.js';
 
 export { declare_resource_structure_panel } from './resource_structure.js';
